Updated all new changes implemented by Shawna. These changes include chaging GetRandomRay in CFG to really be a random ray. Also, c1_towards_c2 shoudlnot be used.  BridgeTest Gauss Maprm obprm have been updated.
Also, Dawens LP updates are here.


git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@1086 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ OBPRM/Stat_Class.h
@@ -120,17 +120,17 @@ protected:
   int NumEdges;
   int NumCC;
 
-  int NumCollDetCalls[MaxCD];
+  unsigned long int NumCollDetCalls[MaxCD];
   char CDNameList[MaxCD][50];
 
   int SizeCC[MaxCC];
 
   char LPNameList[MaxLP][50];
-  int LPConnections[MaxLP];
-  int LPAttempts[MaxLP];
-  int LPCollDetCalls[MaxLP];
+  unsigned long int LPConnections[MaxLP];
+  unsigned long int LPAttempts[MaxLP];
+  unsigned long int LPCollDetCalls[MaxLP];
 
-  std::map<std::string, int> CollDetCountByName;
+  std::map<std::string, unsigned long int> CollDetCountByName;
 
 };
 
@@ -147,7 +147,7 @@ void
 Stat_Class::
 PrintAllStats( Roadmap<CFG, WEIGHT>* rmap, int numCCs) {
   int i;
-  std::map<std::string, int>::const_iterator iter;
+  std::map<std::string, unsigned long int>::const_iterator iter;
   int total=0;
 
   cout << endl << endl << "Local Planners:" << endl;
